# packages/contracts/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateCounterMapNodeDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "breakpoint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateBreakpointDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateCodeSnapshotDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "delivered_prompt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateDeliveredPromptDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateEvidenceDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "execution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateExecutionDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "node_type": {
            "enum": [
              "CLAIM",
              "REASONING",
              "CODE",
              "TEST",
              "QUESTION",
              "RESPONSE",
              "EVIDENCE",
              "BREAKPOINT",
              "ASSISTANCE",
              "MUTATION"
            ],
            "title": "Node Type",
            "type": "string"
          },
          "source_status": {
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ],
            "title": "Source Status",
            "type": "string"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "statement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateTextDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "node_id",
          "node_type",
          "title",
          "summary",
          "stage",
          "source_status"
        ],
        "title": "CandidateCounterMapNodeDetailResponse",
        "type": "object"
      },
      "CandidateCounterMapResponse": {
        "properties": {
          "generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "graph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CounterMapGraph"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "projection_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection Id"
          },
          "projection_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection Version"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "session": {
            "$ref": "#/components/schemas/CounterMapSessionMetadata"
          },
          "status": {
            "enum": [
              "NOT_AVAILABLE",
              "BUILDING",
              "READY",
              "FAILED",
              "STALE"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "session",
          "projection_id",
          "projection_version",
          "schema_version",
          "generated_at",
          "graph",
          "message"
        ],
        "title": "CandidateCounterMapResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "CounterQ application API.",
    "title": "CounterQ API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/countermap/development/fixtures/{fixture_id}/nodes/{node_id}": {
      "get": {
        "operationId": "development_countermap_node_detail_api_countermap_development_fixtures__fixture_id__nodes__node_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "fixture_id",
            "required": true,
            "schema": {
              "title": "Fixture Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateCounterMapNodeDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Development Countermap Node Detail",
        "tags": [
          "countermap"
        ]
      }
    },
    "/api/countermap/development/sessions/{interview_session_id}": {
      "get": {
        "operationId": "development_countermap_status_api_countermap_development_sessions__interview_session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "interview_session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateCounterMapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Development Countermap Status",
        "tags": [
          "countermap"
        ]
      }
    },
    "/api/countermap/development/sessions/{interview_session_id}/inspection": {
      "get": {
        "operationId": "development_countermap_inspection_api_countermap_development_sessions__interview_session_id__inspection_get",
        "parameters": [
          {
            "in": "path",
            "name": "interview_session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevelopmentCounterMapInspection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Development Countermap Inspection",
        "tags": [
          "countermap"
        ]
      }
    },
    "/api/countermap/development/sessions/{interview_session_id}/nodes/{node_id}": {
      "get": {
        "operationId": "development_countermap_node_detail_for_session_api_countermap_development_sessions__interview_session_id__nodes__node_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "interview_session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateCounterMapNodeDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Development Countermap Node Detail For Session",
        "tags": [
          "countermap"
        ]
      }
    },
    "/api/countermap/development/sessions/{interview_session_id}/regenerate": {
      "post": {
        "operationId": "regenerate_countermap_api_countermap_development_sessions__interview_session_id__regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "interview_session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevelopmentCounterMapRegenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevelopmentCounterMapRegenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Countermap",
        "tags": [
          "countermap"
        ]
      }
    },
    "/api/countermap/sessions/{interview_session_id}": {
      "get": {
        "operationId": "candidate_countermap_status_api_countermap_sessions__interview_session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "interview_session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateCounterMapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Candidate Countermap Status",
        "tags": [
          "countermap"
        ]
      }
    },
    "/api/countermap/sessions/{interview_session_id}/nodes/{node_id}": {
      "get": {
        "operationId": "candidate_countermap_node_detail_api_countermap_sessions__interview_session_id__nodes__node_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "interview_session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interview Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateCounterMapNodeDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Candidate Countermap Node Detail",
        "tags": [
          "countermap"
        ]
      }
    },
...
  }
}